feat:编写sku以及spu商品信息以及会员信息编写

+++ product/src/main/java/com/sxx/product/entity/ProductAttrValue.java
@@ -54,7 +54,7 @@ public class ProductAttrValue implements Serializable {
      * 快速展示【是否展示在介绍上；0-否 1-是】
      */
     @TableField(value = "quick_show")
-    private Byte quickShow;
+    private int quickShow;
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
